requirements.txt
# ─── Core Data Science ───────────────────────────────────
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
scipy
matplotlib
seaborn

# ─── Azure SQL Connection ────────────────────────────────
SQLAlchemy
pyodbc

# ─── Environment & Config ────────────────────────────────
python-dotenv
pyyaml
dill

# ─── FastAPI Web App ─────────────────────────────────────
fastapi
uvicorn[standard]
jinja2
python-multipart

# ─── ML Monitoring ───────────────────────────────────────
evidently

# ─── AWS S3 (Model Storage) ──────────────────────────────
boto3

# ─── Project Package ─────────────────────────────────────
setuptools
-e .